requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
transformers==4.43.3
peft==0.11.1
trl==0.9.6
accelerate==0.33.0
bitsandbytes==0.47.0
datasets==2.20.0
einops==0.7.0
sentencepiece==0.2.0
protobuf==4.25.3
evaluate==0.4.2
scipy==1.13.1
tensorboard==2.17.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
gradio==4.44.0